Unit test on secret data remove from output added
